+++ js/main.js
@@ -1,6 +1,6 @@
 import {createRandomAdverts} from './mock.js';
-import {renderMap} from './map.js';
+import {renderMap} from './popup.js';
 
 const similarAdverts = createRandomAdverts();
-renderMap(similarAdverts);
+renderMap(similarAdverts[0]);
